+++ database/data.sql
@@ -1,11 +1,5 @@
--- Use SQL insert statements to add any
--- starting/dummy data to your database tables
-
--- EXAMPLE:
-
---  insert into "todos"
---    ("task", "isCompleted")
---    values
---      ('Learn to code', false),
---      ('Build projects', false),
---      ('Get a job', false);
+insert into "todos" ("task", "isCompleted")
+values
+  ('Review template docs', false),
+  ('Create your first feature', false),
+  ('Ship your MVP', false);